Move the Mxximize functions to Rexl.Solve
