+++ src/Core/Rexl.Solve/Functions/Module.Gen.cs
