+++ src/Core/Rexl.Solve/Functions/Module.cs
