+++ src/Rexl.sln
@@ -59,6 +59,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Apps", "Apps", "{CFD34F94-8
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Benchmark", "Benchmark", "{EC96A595-BEDA-4045-8BA8-C7CCDB5671D4}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Rexl.Bind.TestBase", "Test\Rexl.Bind.TestBase\Rexl.Bind.TestBase.csproj", "{DBEFA827-8D7B-4562-B7D9-D3D116621202}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Rexl.Code.TestBase", "Test\Rexl.Code.TestBase\Rexl.Code.TestBase.csproj", "{E63522ED-3F1C-4324-8FCB-692BE2A5D78E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x64 = Debug|x64
@@ -141,6 +145,14 @@ Global
 		{F254927F-8B17-4A5C-BF93-1BB6D1A87ED2}.Debug|x64.Build.0 = Debug|x64
 		{F254927F-8B17-4A5C-BF93-1BB6D1A87ED2}.Release|x64.ActiveCfg = Release|x64
 		{F254927F-8B17-4A5C-BF93-1BB6D1A87ED2}.Release|x64.Build.0 = Release|x64
+		{DBEFA827-8D7B-4562-B7D9-D3D116621202}.Debug|x64.ActiveCfg = Debug|x64
+		{DBEFA827-8D7B-4562-B7D9-D3D116621202}.Debug|x64.Build.0 = Debug|x64
+		{DBEFA827-8D7B-4562-B7D9-D3D116621202}.Release|x64.ActiveCfg = Release|x64
+		{DBEFA827-8D7B-4562-B7D9-D3D116621202}.Release|x64.Build.0 = Release|x64
+		{E63522ED-3F1C-4324-8FCB-692BE2A5D78E}.Debug|x64.ActiveCfg = Debug|x64
+		{E63522ED-3F1C-4324-8FCB-692BE2A5D78E}.Debug|x64.Build.0 = Debug|x64
+		{E63522ED-3F1C-4324-8FCB-692BE2A5D78E}.Release|x64.ActiveCfg = Release|x64
+		{E63522ED-3F1C-4324-8FCB-692BE2A5D78E}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -166,6 +178,8 @@ Global
 		{C91BE169-F49D-4EF6-8A45-55DD3C9F31D0} = {6ECE0CC4-5B0C-476A-8B98-46C6B7E6816D}
 		{F254927F-8B17-4A5C-BF93-1BB6D1A87ED2} = {EC96A595-BEDA-4045-8BA8-C7CCDB5671D4}
 		{8C79E433-5689-41DB-8175-1E5CDCDE8B10} = {CFD34F94-8B32-4961-B0BD-8BA72230FFCE}
+		{DBEFA827-8D7B-4562-B7D9-D3D116621202} = {6ECE0CC4-5B0C-476A-8B98-46C6B7E6816D}
+		{E63522ED-3F1C-4324-8FCB-692BE2A5D78E} = {6ECE0CC4-5B0C-476A-8B98-46C6B7E6816D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {B4DF49D6-B259-43EF-AC43-27CC8D6E4976}
