+++ src/Test/Baseline/Bind/Binder/With/ReducePull.txt
@@ -1,4 +1,4 @@
-**** New globals: {A:i8, B:i8, C:i8, M:M{param P:i8, msr S:i8}, QA:i8?, T:{A:i8}, Txt:s, b:b, si8:i8*, si8_2:i8*}
+**** New globals: {A:i8, B:i8, C:i8, QA:i8?, T:{A:i8}, Txt:s, b:b, si8:i8*, si8_2:i8*}
 
 > ForEach(With (s: Wrap(si8), s ++ s), it^3)
 ForEach(With(s : Wrap(si8), s ++ s), it^3) : i8*
