+++ src/Test/Baseline/Bind/FuncSig/All.txt
@@ -849,9 +849,6 @@ Flags: __, min: 2, max: 2, Path: Mod
   Computes the remainder when 'x' is divided by 'y' with the quotient rounded toward zero to the nearest integer value. Note that the sign of the result is the same as the sign of 'x'.
     0) x: The numerator to be divided by denominator.
     1) y: The denominator by which numerator is to be divided.
-Flags: __, min: 2, max: 3, Path: Module.Maximize
-Flags: __, min: 2, max: 3, Path: Module.Minimize
-Flags: __, min: 3, max: 4, Path: Module.Optimize
 Flags: __, min: 1, max: 1, Path: Null
 Flags: __, min: 1, max: 1, Path: NullItem
 Flags: __, min: 1, max: 1, Path: Opt
