+++ src/Test/Rexl.Bind.Test/Rexl.Bind.Test.csproj
@@ -7,11 +7,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="..\..\Shared.Test\BoundTreeValidator.cs" Link="Shared\BoundTreeValidator.cs" />
-    <Compile Include="..\..\Shared.Test\RexlLineTestsBase.cs" Link="Shared\RexlLineTestsBase.cs" />
-    <Compile Include="..\..\Shared.Test\RexlTestsBase.cs" Link="Shared\RexlTestsBase.cs" />
-    <Compile Include="..\..\Shared.Test\TestOperations.cs" Link="Shared\TestOperations.cs" />
-    <Compile Include="..\..\Shared.Test\TestProcedures.cs" Link="Shared\TestProcedures.cs" />
     <Compile Include="..\..\Shared\Validation.cs" Link="Shared\Validation.cs" />
   </ItemGroup>
 
@@ -27,6 +22,7 @@
   <ItemGroup>
     <ProjectReference Include="..\..\Core\Rexl.Base\Rexl.Base.csproj" />
     <ProjectReference Include="..\..\Core\Rexl.Bind\Rexl.Bind.csproj" />
+    <ProjectReference Include="..\Rexl.Bind.TestBase\Rexl.Bind.TestBase.csproj" />
   </ItemGroup>
 
   <ItemGroup>
