+++ src/Test/Rexl.Bind.Test/Scripts/Binder/General/Module.txt
@@ -6,8 +6,11 @@ M.A
 M.X
 M
 M => { A: 17 }
+
+// Unknown function.
 Module.Maximize(M, S)
 M->Maximize(S)
+
 R
 R.A
 R.M
