+++ src/Test/Rexl.Bind.Test/Scripts/Binder/Module/Basic.txt
@@ -142,40 +142,3 @@ M=>("hi")
 R with { B:"hi" }
 ###
 R => { A:17 }
-###
-Module.Maximize(M, S)
-###
-M->Minimize(S)
-###
-QM->Minimize(S)
-###
-Module.Optimize(M, S, b)
-###
-QM->Optimize(S, b)
-###
-Module.Maximize(M, s) // Error.
-###
-Module.Maximize(M, A) // Error.
-###
-M->Maximize(X) // Error.
-###
-Module.Maximize(R, A) // Error.
-###
-Module.Maximize(QR, A) // Error.
-###
-M->Maximize(S, "GLPK")
-###
-Module.Maximize(M, 0) // Error.
-###
-Module.Maximize(M, 2 * S) // Error.
-###
-Module.Minimize(M, -S) // Error.
-###
-Module.Minimize(M, T) // Error.
-###
-M->Optimize(S, A > 10) // Error.
-###
-M->Optimize(S, b, "highs")
-###
-// Hoisting out of optimize expression.
-Range(10)->ForEach(as n, Module.Maximize(M with { A : R.A * R.A * n }, S))
