+++ src/Test/Rexl.Bind.TestBase/BoundTreeValidator.cs
