+++ src/Test/Rexl.Bind.TestBase/Rexl.Bind.TestBase.csproj
@@ -4,16 +4,12 @@
     <TargetFramework>net6.0</TargetFramework>
     <IsPackable>false</IsPackable>
     <IsTestProject>true</IsTestProject>
-    <DefineConstants>$(DefineConstants);WITH_CODE</DefineConstants>
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Include="..\..\Shared.Test\RexlLineTestsBase.cs" Link="Shared\RexlLineTestsBase.cs" />
+    <Compile Include="..\..\Shared.Test\RexlTestsBase.cs" Link="Shared\RexlTestsBase.cs" />
     <Compile Include="..\..\Shared\Validation.cs" Link="Shared\Validation.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="..\Baseline\Flow\**\*" LinkBase="Test\Baseline" />
-    <None Include="..\XOutput\Flow\**\*" LinkBase="Test\XOutput" />
   </ItemGroup>
 
   <ItemGroup>
@@ -23,10 +19,6 @@
   <ItemGroup>
     <ProjectReference Include="..\..\Core\Rexl.Base\Rexl.Base.csproj" />
     <ProjectReference Include="..\..\Core\Rexl.Bind\Rexl.Bind.csproj" />
-    <ProjectReference Include="..\..\Core\Rexl.Code\Rexl.Code.csproj" />
-    <ProjectReference Include="..\..\Core\Rexl.Flow\Rexl.Flow.csproj" />
-    <ProjectReference Include="..\..\Core\Rexl.Harness\Rexl.Harness.csproj" />
-    <ProjectReference Include="..\Rexl.Code.TestBase\Rexl.Code.TestBase.csproj" />
   </ItemGroup>
 
 </Project>
