+++ src/Test/Rexl.Code.Test/BlockTests.cs
@@ -3,6 +3,7 @@
 
 using System.Text;
 using System.Threading.Tasks;
+using Microsoft.Rexl;
 using Microsoft.Rexl.Code;
 using Microsoft.Rexl.Lex;
 
@@ -13,6 +14,19 @@ namespace RexlTest;
 [TestClass]
 public sealed class BlockTests : BlockTestsBase<bool>
 {
+    private readonly OperationRegistry _opers;
+    private readonly GeneratorRegistry _gens;
+
+    protected override OperationRegistry Operations => _opers;
+    protected override GeneratorRegistry Generators => _gens;
+
+    public BlockTests()
+        : base()
+    {
+        _opers = new AggregateOperationRegistry(TestFunctions.Instance, MultiFormOperations.Instance);
+        _gens = new AggregateGeneratorRegistry(TestGenerators.Instance, MultiFormGenerators.Instance);
+    }
+
     [TestMethod]
     public async Task SegmentedTests()
     {
