+++ src/Test/Rexl.Code.Test/Rexl.Code.Test.csproj
@@ -8,18 +8,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="..\..\Shared.Test\BlockTestsBase.cs" Link="Shared\BlockTestsBase.cs" />
-    <Compile Include="..\..\Shared.Test\BoundTreeValidator.cs" Link="Shared\BoundTreeValidator.cs" />
-    <Compile Include="..\..\Shared.Test\RexlLineTestsBase.cs" Link="Shared\RexlLineTestsBase.cs" />
-    <Compile Include="..\..\Shared.Test\RexlTestsBase.cs" Link="Shared\RexlTestsBase.cs" />
-    <Compile Include="..\..\Shared.Test\RexlTestsBaseType.cs" Link="Shared\RexlTestsBaseType.cs" />
-    <Compile Include="..\..\Shared.Test\TestOperations.cs" Link="Shared\TestOperations.cs" />
-    <Compile Include="..\..\Shared.Test\TestOperations.Gen.cs" Link="Shared\TestOperations.Gen.cs" />
-    <Compile Include="..\..\Shared.Test\TestProcedures.cs" Link="Shared\TestProcedures.cs" />
-    <Compile Include="..\..\Shared.Test\TestProcedures.Gen.cs" Link="Shared\TestProcedures.Gen.cs" />
-    <Compile Include="..\..\Shared.Test\TestTypeManager.cs" Link="Shared\TestTypeManager.cs" />
-    <Compile Include="..\..\Shared.Test\TestTypeManager.Impls.cs" Link="Shared\TestTypeManager.Impls.cs" />
-    <Compile Include="..\..\Shared.Test\TestUtils.cs" Link="Shared\TestUtils.cs" />
     <Compile Include="..\..\Shared\Validation.cs" Link="Shared\Validation.cs" />
   </ItemGroup>
 
@@ -38,6 +26,7 @@
     <ProjectReference Include="..\..\Core\Rexl.Code\Rexl.Code.csproj" />
     <ProjectReference Include="..\..\Core\Rexl.Flow\Rexl.Flow.csproj" />
     <ProjectReference Include="..\..\Core\Rexl.Harness\Rexl.Harness.csproj" />
+    <ProjectReference Include="..\Rexl.Code.TestBase\Rexl.Code.TestBase.csproj" />
   </ItemGroup>
 
   <ItemGroup>
