--- src/Test/Rexl.Code.Test/Scripts/CodeGen/Special/CodeGenHost.txt
@@ -1,3 +1,0 @@
-﻿WrapCallSvcFromHost(1)
-WrapCallSvcFromHost(3) + WrapCallSvcFromHost(5)
-Sum(Range(3), WrapCallSvcFromHost(it))
