--- src/Test/Rexl.Code.Test/Scripts/Module/Optimize/Basic.txt
@@ -1,13 +1,0 @@
-P := plan {
-    var X from 0.0 to 1.0;
-    var Y from 0.0 to 1.0;
-
-    msr M := 2 * X + Y;
-
-    con Limit := X + Y <= 1.5;
-};
-
-P;
-
-"*** Maximize ***";
-P->Maximize(M); // Fails since solvers aren't imported.
