+++ src/Test/Rexl.Code.TestBase/Rexl.Code.TestBase.csproj
@@ -4,16 +4,13 @@
     <TargetFramework>net6.0</TargetFramework>
     <IsPackable>false</IsPackable>
     <IsTestProject>true</IsTestProject>
-    <DefineConstants>$(DefineConstants);WITH_CODE;WITH_ONNX</DefineConstants>
+    <DefineConstants>$(DefineConstants);WITH_CODE</DefineConstants>
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Include="..\..\Shared.Test\BlockTestsBase.cs" Link="Shared\BlockTestsBase.cs" />
+    <Compile Include="..\..\Shared.Test\RexlTestsBaseType.cs" Link="Shared\RexlTestsBaseType.cs" />
     <Compile Include="..\..\Shared\Validation.cs" Link="Shared\Validation.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="..\Baseline\Onnx\**\*" LinkBase="Test\Baseline" />
-    <None Include="..\XOutput\Onnx\**\*" LinkBase="Test\XOutput" />
   </ItemGroup>
 
   <ItemGroup>
@@ -25,9 +22,8 @@
     <ProjectReference Include="..\..\Core\Rexl.Bind\Rexl.Bind.csproj" />
     <ProjectReference Include="..\..\Core\Rexl.Code\Rexl.Code.csproj" />
     <ProjectReference Include="..\..\Core\Rexl.Flow\Rexl.Flow.csproj" />
-    <ProjectReference Include="..\..\Core\Rexl.Onnx\Rexl.Onnx.csproj" />
     <ProjectReference Include="..\..\Core\Rexl.Harness\Rexl.Harness.csproj" />
-    <ProjectReference Include="..\Rexl.Code.TestBase\Rexl.Code.TestBase.csproj" />
+    <ProjectReference Include="..\Rexl.Bind.TestBase\Rexl.Bind.TestBase.csproj" />
   </ItemGroup>
 
 </Project>
