+++ src/Test/Rexl.Code.TestBase/TestTypeManager.Impls.cs
