+++ src/Test/Rexl.Code.TestBase/TestTypeManager.cs
