+++ src/Test/Rexl.Flow.Test/GridTests.cs
@@ -2635,7 +2635,7 @@ partial class DocumentTests
                 return false;
             }
 
-            info = TestOperations.Instance.GetInfo(name);
+            info = TestFunctions.Instance.GetInfo(name);
             Validation.Assert(info is null || info.Oper is not null);
             return info != null;
         }
