+++ src/Test/Rexl.Solve.Test/Rexl.Solve.Test.csproj
@@ -8,16 +8,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="..\..\Shared.Test\BlockTestsBase.cs" Link="Shared\BlockTestsBase.cs" />
-    <Compile Include="..\..\Shared.Test\BoundTreeValidator.cs" Link="Shared\BoundTreeValidator.cs" />
-    <Compile Include="..\..\Shared.Test\RexlTestsBase.cs" Link="Shared\RexlTestsBase.cs" />
-    <Compile Include="..\..\Shared.Test\RexlTestsBaseType.cs" Link="Shared\RexlTestsBaseType.cs" />
-    <Compile Include="..\..\Shared.Test\TestOperations.cs" Link="Shared\TestOperations.cs" />
-    <Compile Include="..\..\Shared.Test\TestOperations.Gen.cs" Link="Shared\TestOperations.Gen.cs" />
-    <Compile Include="..\..\Shared.Test\TestProcedures.cs" Link="Shared\TestProcedures.cs" />
-    <Compile Include="..\..\Shared.Test\TestProcedures.Gen.cs" Link="Shared\TestProcedures.Gen.cs" />
-    <Compile Include="..\..\Shared.Test\TestTypeManager.cs" Link="Shared\TestTypeManager.cs" />
-    <Compile Include="..\..\Shared.Test\TestTypeManager.Impls.cs" Link="Shared\TestTypeManager.Impls.cs" />
     <Compile Include="..\..\Shared\Validation.cs" Link="Shared\Validation.cs" />
   </ItemGroup>
 
@@ -36,6 +26,8 @@
     <ProjectReference Include="..\..\Core\Rexl.Code\Rexl.Code.csproj" />
     <ProjectReference Include="..\..\Core\Rexl.Solve\Rexl.Solve.csproj" />
     <ProjectReference Include="..\..\Core\Rexl.Harness\Rexl.Harness.csproj" />
+    <ProjectReference Include="..\Rexl.Bind.TestBase\Rexl.Bind.TestBase.csproj" />
+    <ProjectReference Include="..\Rexl.Code.TestBase\Rexl.Code.TestBase.csproj" />
   </ItemGroup>
 
 </Project>
